changes based on prev feedback
