+++ src/components/Strategies/Blocks/index.ts
@@ -3,7 +3,7 @@ import { default as Arbitrum_Bridgeworld_Claim } from './Arbitrum_Bridgeworld_Cl
 import { default as Arbitrum_Magic_Send } from './Arbitrum_Magic_Send';
 
 export { default as Repeat } from './Repeat';
-export { default as SigningPopup } from './SigningPopup';
+export { default as SigningPopup } from '../SigningPopup';
 
 const blockConfig: {
   [key in StrategyBlock]: { component: () => JSX.Element; requiresFallback: boolean };
