+++ src/utils.ts
@@ -83,8 +83,8 @@ export function formatCurrency(value: number, decimals = 2): string {
   return formatter.format(value || 0);
 }
 
-export function calculatePercent(value: number, maxvalue: number): number {
-  return Math.round((value / maxvalue) * 100);
+export function calculatePercent(value: number, maxValue: number): number {
+  return Math.round((value / maxValue) * 100);
 }
 
 export function isEmptyName(value: string): boolean {
